+++ src/Constants.ts
@@ -1,5 +1,5 @@
 export class Constants {
-    public static readonly LOG_ID: string = 'log';
+    public static readonly LogLanguageId: string = 'log';
 
     public static readonly ContextNameIsShowingProgressIndicators = 'logFileHighlighter.isShowingProgressIndicators';
 }