+++ src/TailController.ts
@@ -75,7 +75,7 @@ export class TailController {
         const textEditor = vscode.window.activeTextEditor;
         const visibleRanges = vscode.window.activeTextEditor.visibleRanges
 
-        if (textEditor?.document.languageId === Constants.LOG_ID) {
+        if (textEditor?.document.languageId === Constants.LogLanguageId) {
             const lastLine = textEditor.document.lineCount - 1;
             const lastVisibleRange = visibleRanges[0].end.line;