+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
   To disable this feature altogether, set the `logFileHighlighter.enableTailMode` setting to `false`.
 
+* Also fixed multiple bugs with custom patterns that occurred when a log file was modified, either externally (new log lines being added at the end) or interactive editing of log files (which is an edge case for sure, as log files are generally not modified). The issue was out-of-sync pattern positions.
+
 ### 3.1.0 - 22 Juni 2024
 
 **New feature**